+++ opnsense_cli/api/exceptions.py
